requirements.txt
Django==2.1.9
djangorestframework==3.10.2
pyodbc==4.0.26
django-pyodbc-azure==2.1.0.0
meraki==1.3.0